inheritance exercises continued. Can't get privatelyOwned method to work correctly for some reason

+++ inheritance/testing-inheritance/exercises/src/main/java/transportation/Boat.java
@@ -5,7 +5,7 @@ public class Boat extends Vehicle{
     private String make;
     private String color;
 
-    public Boat(String type, boolean hasWheels, double occupancy, boolean isPrivatelyOwned) {
+    public Boat(String type, boolean hasWheels, int occupancy, boolean isPrivatelyOwned) {
         super(type, hasWheels, occupancy, isPrivatelyOwned);
     }
