+++ inheritance/testing-inheritance/exercises/src/main/java/transportation/Plane.java
@@ -4,7 +4,7 @@ public class Plane extends Vehicle{
 
     private double topSpeed;
 
-    public Plane(String type, boolean hasWheels, double occupancy, boolean isPrivatelyOwned) {
+    public Plane(String type, boolean hasWheels, int occupancy, boolean isPrivatelyOwned) {
         super(type, hasWheels, occupancy, isPrivatelyOwned);
     }
